loss error

+++ utils/loss_Ncla.py
@@ -189,9 +189,9 @@ class ComputeLoss:
         tcls, tbox, tkpt, indices, anch = [], [], [], [], []
         if self.kpt_label:
             # gain = torch.ones(41, device=targets.device)  # normalized to gridspace gain
-            gain = torch.ones(self.kpt_num*2 +7 , device=targets.device)  # normalized to gridspace gain
+            gain = torch.ones(self.kpt_num*2 +7 , device=targets.device).long()  # normalized to gridspace gain
         else:
-            gain = torch.ones(7, device=targets.device)  # normalized to gridspace gain
+            gain = torch.ones(7, device=targets.device).long()  # normalized to gridspace gain
         ai = torch.arange(na, device=targets.device).float().view(na, 1).repeat(1, nt)  # same as .repeat_interleave(nt)
         targets = torch.cat((targets.repeat(na, 1, 1), ai[:, :, None]), 2)  # append anchor indices
